fixed: String handling issues in 64-bit

+++ indra/newview/llsechandler_basic.cpp
@@ -738,7 +738,7 @@ bool _cert_subdomain_wildcard_match(const std::string& subdomain,
 {
 	// split wildcard into the portion before the *, and the portion after
 
-	int wildcard_pos = wildcard.find_first_of('*');	
+	size_t wildcard_pos = wildcard.find_first_of('*');	
 	// check the case where there is no wildcard.
 	if(wildcard_pos == wildcard.npos)
 	{
@@ -770,7 +770,7 @@ bool _cert_subdomain_wildcard_match(const std::string& subdomain,
 	std::string new_subdomain = subdomain.substr(wildcard_pos, subdomain.npos);
 	
 	// iterate through the current subdomain, finding instances of the match string.
-	int sub_pos = new_subdomain.find_first_of(new_wildcard_match_string);
+	size_t sub_pos = new_subdomain.find_first_of(new_wildcard_match_string);
 	while(sub_pos != std::string::npos)
 	{
 		new_subdomain = new_subdomain.substr(sub_pos, std::string::npos);
@@ -802,8 +802,8 @@ bool _cert_hostname_wildcard_match(const std::string& hostname, const std::strin
 	std::string new_cn = common_name;
 	
 	// find the last '.' in the hostname and the match name.
-	int subdomain_pos = new_hostname.find_last_of('.');
-	int subcn_pos = new_cn.find_last_of('.');
+	size_t subdomain_pos = new_hostname.find_last_of('.');
+	size_t subcn_pos = new_cn.find_last_of('.');
 	
 	// if the last char is a '.', strip it
 	if(subdomain_pos == (new_hostname.length()-1))
